clustermesh: helm: pass cilium-host-aliases in etcd config
Add the cilium-host-aliases in the etcd config which should be fully replacing
hostaliases in Cilium 1.21. We are keeping hostAliases for now to keep
downgrade safe (secrets updated without cilium-clustermesh-apiserver-host and
potentially without hostAliases if we were already removing it here).

We are force disabling this in the conformance ClusterMesh CI via a non
documented helm values to already exercise that this is working properly
while the conformance upgrade CI stays on default setting to keep the
existing flow tested!

Signed-off-by: Arthur Outhenin-Chalandre <git@mrfreezeex.fr>